backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
yfinance>=0.2.40
anthropic>=0.40.0
scikit-learn>=1.5.0
xgboost>=2.1.0
shap>=0.46.0
numpy>=1.26.0
pandas>=2.2.0
httpx>=0.27.0
vaderSentiment>=3.3.1
python-multipart>=0.0.12
a2a-sdk[http-server]>=0.3.0
mcp>=1.0.0,<2.0.0
claude-agent-sdk>=0.1.53
sentry-sdk[starlette]>=2.0.0
cryptography>=42.0.0
slowapi>=0.1.9
limits>=3.12.0
